reviewing first implementation

+++ regime_inde/regime_inde/src/main/java/fr/sebaseg/regime_inde/repository/CfpRateRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CfpRateRepository extends CrudRepository<CfpRate, Integer> {
+public interface CfpRateRepository extends CrudRepository<CfpRate, Long> {
 }
